+++ changelog.md
@@ -5,6 +5,7 @@
 ## next
 - Fixed non-ascii characters are not rendered correctly, json.dumps is applied to user texts and json.loads is applied to platform texts
 - Reversed the order of the buttons of the sensitive question
+- Reminded to the user the attributes he chose when accepting the question
 
 ## 2.0.2
 - Updated translations